+++ archive/v15.0-commercial/wvs/__init__.py
@@ -1,3 +1,0 @@
-"""WVS - Web Vulnerability Scanner"""
-__version__ = "1.0.0"
-__author__ = "Security Team"