+++ archive/v15.0/src/wvs/modules/report/__init__.py
@@ -1,4 +1,0 @@
-"""WVS v18.2 - Report Generator Module"""
-from .report_generator import ReportGenerator, ScanReport
-
-__all__ = ["ReportGenerator", "ScanReport"]