+++ archive/v15.0/src/wvs/modules/verifier/__init__.py
@@ -1,4 +1,0 @@
-"""WVS v18 - Vulnerability Verifier Module"""
-from .vuln_verifier import VulnerabilityVerifier, VerificationResult
-
-__all__ = ["VulnerabilityVerifier", "VerificationResult"]