+++ archive/v15.0/src/wvs/plugins/compliance/__init__.py
@@ -1,2 +1,0 @@
-"""WVS v18.0 - Compliance Plugins"""
-from .scanner import ComplianceScanner