+++ archive/v17/wvs/integrations/__init__.py
@@ -1,5 +1,0 @@
-"""WVS v17.0 - 集成模块"""
-
-from .manager import IntegrationManager, IntegrationConfig, IntegrationType
-
-__all__ = ["IntegrationManager", "IntegrationConfig", "IntegrationType"]