+++ archive/v18.4/src/wvs/modules/__init__.py
@@ -1,1 +1,0 @@
-"""WVS v18 modules"""