+++ archive/v18.4/src/wvs/plugins/__init__.py
@@ -1,1 +1,0 @@
-"""WVS v18.0 - Cloud Native Security Plugins"""