+++ archive/v18.4/src/wvs/plugins/cloud/__init__.py
@@ -1,2 +1,0 @@
-"""WVS v18.0 - Cloud Scanner"""
-from .scanner import CloudNativeScanner