+++ archive/v18.4/src/wvs/plugins/mobile/__init__.py
@@ -1,2 +1,0 @@
-"""WVS v18.0 - Mobile Security Plugins"""
-from .scanner import MobileScanner