+++ archive/v18.4/src/wvs/report/__init__.py
@@ -1,4 +1,0 @@
-"""WVS v18 - Report Module"""
-from .report_generator import ReportGenerator, ScanReporter, VulnerabilityStats
-
-__all__ = ['ReportGenerator', 'ScanReporter', 'VulnerabilityStats']