+++ archive/v18/wvs/modules/lfi/__init__.py
@@ -1,4 +1,0 @@
-"""LFI + Log Poisoning RCE Scanner"""
-from .lfi_scanner import LFIScanner
-from .log_poisoning import LogPoisoningDetector
-from .phpliteadmin_scanner import PHPLiteAdminScanner, PHPLiteAdminResult, scan_phpliteadmin